Fix compile errors:
* Warning over conversion from size_t to uint32_t
* Use of variable inside GENERATE expression - see
https://github.com/catchorg/Catch2/blob/devel/docs/generators.md

+++ driver/support_library/src/WeightEncoder.cpp
@@ -1779,7 +1779,7 @@ EncodedWeights WeightEncoder::Encode(const TensorInfo& weightsTensorInfo,
     for (size_t og = 0; og < numOfmInParallel; ++og)
     {
         waitHandles[og] = std::async(
-            [&](uint32_t og) {
+            [&](size_t og) {
                 for (uint32_t ofm : perOgOfms[og])
                 {
                     const uint32_t iteration = ofm % numIterationsOfm;
